Updated admin panel to deal with employee_type variable. Added JCB to add employee.

+++ src/database/DBConnection.java
@@ -86,8 +86,8 @@ public class DBConnection {
                 Double onHoliday = resultSet.getDouble("onHoliday");
                 Double offSick = resultSet.getDouble("offSick");
                 int ward_ID = resultSet.getInt("ward_ID");
-
                 String password = resultSet.getString("password");
+                String employee_type = resultSet.getString("employee_type");
                 String privilege = resultSet.getString("privilege");
 
                 Employee employee = new Employee();
@@ -114,6 +114,7 @@ public class DBConnection {
                 employee.setOffSick(offSick);
                 employee.setWard_ID(ward_ID);
                 employee.setPassword(password);
+                employee.setEmployee_type(employee_type);
                 employee.setPrivilege(privilege);
 
                 employeeList.add(employee);
@@ -130,13 +131,13 @@ public class DBConnection {
     }
 
     public void createEmployee(String fNameIn, String sNameIn, Calendar DOBIn, String contactNumIn, String emailIn,
-                               double numHolidaysIn, double contractHoursIn, double salary, int ward_IDIn, String password, String privilege) {
+                               double numHolidaysIn, double contractHoursIn, double salary, int ward_IDIn, String password, String privilege, String employee_type) {
 
         getDBConnection();
 
         try {
             PreparedStatement preparedStatement;
-            preparedStatement = connection.prepareStatement("INSERT INTO employee VALUES(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);");
+            preparedStatement = connection.prepareStatement("INSERT INTO employee VALUES(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);");
             preparedStatement.setInt(1, 0); // employee id 0 as it's auto incremented in DB
             preparedStatement.setString(2, fNameIn);
             preparedStatement.setString(3, sNameIn);
@@ -155,7 +156,7 @@ public class DBConnection {
             preparedStatement.setInt(13, ward_IDIn);
             preparedStatement.setString(14, password);
             preparedStatement.setString(15, privilege);
-//            System.out.print("Details; \n" + fNameIn + " " + sNameIn + " " + sqlDate.toString() + " " + contactNumIn + " " + emailIn + " " + numHolidaysIn + " " + contractHoursIn + " " + salary + " " + ward_IDIn);
+            preparedStatement.setString(16, employee_type);
             preparedStatement.executeUpdate();
 
         } catch (Exception e) {
@@ -168,41 +169,55 @@ public class DBConnection {
         }
     }
 
-    public void removeEmployee(Employee employee) {
+    public boolean removeEmployee(Employee employee) {
+        boolean shiftsAssigned = false;
+        shiftsAssigned = removeEmployeeFromShift_Employee(employee);
+        boolean employeeRemoved = false;
 
-        removeEmployeeFromShift_Employee(employee);
+        if(!shiftsAssigned) {
+            getDBConnection();
 
-        getDBConnection();
+            try {
+                String name = employee.getfName();
+                PreparedStatement preparedStatement;
+                preparedStatement = connection.prepareStatement("DELETE from employee WHERE emp_ID = ?;");
+                preparedStatement.setInt(1, employee.getEmp_ID());
+                int count = preparedStatement.executeUpdate();
 
-        try {
-            String name = employee.getfName();
-            PreparedStatement preparedStatement;
-            preparedStatement = connection.prepareStatement("DELETE from employee WHERE emp_ID = ?;");
-            preparedStatement.setInt(1, employee.getEmp_ID());
-            int count = preparedStatement.executeUpdate();
+                if (count > 0) {
+                    JOptionPane.showMessageDialog(null, name + " has been removed.", "Success", JOptionPane.INFORMATION_MESSAGE);
+                    employeeRemoved = true;
+                }
 
-            if (count > 0)
-                JOptionPane.showMessageDialog(null, name + " has been removed.", "Success", JOptionPane.INFORMATION_MESSAGE);
-            else
-                JOptionPane.showMessageDialog(null, "Employee not removed.", "Error", JOptionPane.ERROR_MESSAGE);
-
-        } catch (Exception e) {
-            e.getStackTrace();
-        } finally {
-            closeResultSet();
-            closeStatement();
-            closeConnection();
+            } catch (Exception e) {
+                e.getStackTrace();
+            } finally {
+                closeResultSet();
+                closeStatement();
+                closeConnection();
+            }
         }
+
+        return employeeRemoved;
+
     }
 
-    private void removeEmployeeFromShift_Employee(Employee employee) {
+    private boolean removeEmployeeFromShift_Employee(Employee employee) {
+
+        boolean shiftsAssigned = false;
+
         getDBConnection();
 
         try {
             PreparedStatement preparedStatement;
-            preparedStatement = connection.prepareStatement("DELETE from shift_employee WHERE emp_ID = ?;");
+            preparedStatement = connection.prepareStatement("SELECT * from shift_employee WHERE emp_ID = ?;");
             preparedStatement.setInt(1, employee.getEmp_ID());
-            preparedStatement.executeUpdate();
+            int results = preparedStatement.executeUpdate();
+
+            if (results > 0)
+                shiftsAssigned = true;
+            else
+                shiftsAssigned = false;
 
         } catch (Exception e) {
             e.getStackTrace();
@@ -211,6 +226,8 @@ public class DBConnection {
             closeStatement();
             closeConnection();
         }
+
+        return shiftsAssigned;
     }
 
     public void updateEmployee(Employee employee) {
@@ -221,7 +238,7 @@ public class DBConnection {
             PreparedStatement preparedStatement;
             preparedStatement = connection.prepareStatement("UPDATE employee SET fName = ?, sName = ?, " +
                     "DOB = ?, contactNum = ?, email = ?, numHolidays = ?, contractHours = ?, " +
-                    "salary = ?, onHoliday = ?, offSick = ?, lastShift = ?, ward_ID = ?, password = ?, privilege = ? WHERE emp_ID = " + employee.getEmp_ID() + ";");
+                    "salary = ?, onHoliday = ?, offSick = ?, lastShift = ?, ward_ID = ?, password = ?, privilege = ?, employee_type = ? WHERE emp_ID = " + employee.getEmp_ID() + ";");
             preparedStatement.setString(1, employee.getfName());
             preparedStatement.setString(2, employee.getsName());
 
@@ -239,6 +256,7 @@ public class DBConnection {
             preparedStatement.setInt(12, employee.getWard_ID());
             preparedStatement.setString(13, employee.getPassword());
             preparedStatement.setString(14, employee.getPrivilege());
+            preparedStatement.setString(15, employee.getEmployee_type());
             int count = preparedStatement.executeUpdate();
 
             if (count > 0)
